refactor(gui): break down model config components
