+++ extensions/vscode/package.json
@@ -1,7 +1,7 @@
 {
   "name": "continue",
   "icon": "media/icon.png",
-  "version": "0.9.163",
+  "version": "0.9.164",
   "repository": {
     "type": "git",
     "url": "https://github.com/continuedev/continue"
