+++ gui/src/components/modelSelection/ModelCard.tsx
@@ -7,13 +7,27 @@ import {
   lightGray,
   vscFocusBorder,
 } from "..";
-import {
-  MODEL_PROVIDER_TAG_COLORS,
-  PROVIDER_INFO,
-  PackageDimension,
-} from "../../util/modelData";
+import { PROVIDER_INFO } from "../../pages/modelConfig/modelProviders";
 import HeaderButtonWithText from "../HeaderButtonWithText";
 import InfoHover from "../InfoHover";
+import { PackageDimension } from "../../pages/modelconfig/modelPackages";
+import ModelProviderTag, { ModelProviderTagVals } from "./ModelProviderTag";
+
+interface ModelCardProps {
+  title: string;
+  description: string;
+  tags?: ModelProviderTagVals[];
+  refUrl?: string;
+  icon?: string;
+  onClick?: (
+    e: React.MouseEvent<HTMLDivElement, MouseEvent>,
+    dimensionChoices?: string[],
+    selectedProvider?: string,
+  ) => void;
+  disabled?: boolean;
+  dimensions?: PackageDimension[];
+  providerOptions?: string[];
+}
 
 const Div = styled.div<{ color: string; disabled: boolean; hovered: boolean }>`
   border: 1px solid ${lightGray};
@@ -71,31 +85,15 @@ const DimensionOptionDiv = styled.div<{ selected: boolean }>`
   }
 `;
 
-interface ModelCardProps {
-  title: string;
-  description: string;
-  tags?: string[];
-  refUrl?: string;
-  icon?: string;
-  onClick?: (
-    e: React.MouseEvent<HTMLDivElement, MouseEvent>,
-    dimensionChoices?: string[],
-    selectedProvider?: string
-  ) => void;
-  disabled?: boolean;
-  dimensions?: PackageDimension[];
-  providerOptions?: string[];
-}
-
 function ModelCard(props: ModelCardProps) {
   const [dimensionChoices, setDimensionChoices] = useState<string[]>(
-    props.dimensions?.map((d) => Object.keys(d.options)[0]) || []
+    props.dimensions?.map((d) => Object.keys(d.options)[0]) || [],
   );
 
   const [hovered, setHovered] = useState(false);
 
   const [selectedProvider, setSelectedProvider] = useState<string | undefined>(
-    undefined
+    undefined,
   );
 
   useEffect(() => {
@@ -136,21 +134,11 @@ function ModelCard(props: ModelCardProps) {
           )}
           <h3>{props.title}</h3>
         </div>
-        {props.tags?.map((tag) => {
-          return (
-            <span
-              style={{
-                backgroundColor: `${MODEL_PROVIDER_TAG_COLORS[tag]}55`,
-                color: "white",
-                padding: "2px 4px",
-                borderRadius: defaultBorderRadius,
-                marginRight: "4px",
-              }}
-            >
-              {tag}
-            </span>
-          );
-        })}
+
+        {props.tags?.map((tag, i) => (
+          <ModelProviderTag key={i} tag={tag} />
+        ))}
+
         <p>{props.description}</p>
 
         {props.refUrl && (
