+++ gui/src/pages/modelConfig/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./ModelConfig";
