+++ gui/src/util/index.ts
@@ -1,4 +1,5 @@
 import { getLocalStorage } from "./localStorage";
+import _ from "lodash";
 
 type Platform = "mac" | "linux" | "windows" | "unknown";
 
@@ -74,4 +75,32 @@ export function isPrerelease() {
     return true;
   }
   return false;
+}
+
+/**
+ * Updates the values of an object's properties based on the specified paths.
+ *
+ * This function creates a deep clone of the provided object and updates its properties
+ * based on the given path-to-value mappings. If a value in the mapping is a function,
+ * it will be used to determine the new value for the property. Otherwise, the value itself
+ * will be set for the property.
+ *
+ * @param {Object} old - The original object to be cloned and updated.
+ * @param {Object} pathToValue - An object where the keys represent the paths to the properties
+ *                               to be updated and the values represent the new values or functions
+ *                               to determine the new values.
+ * @returns {Object} A new object with the updated values.
+ */
+export function updatedObj(old: any, pathToValue: { [key: string]: any }) {
+  const newObject = _.cloneDeep(old);
+
+  for (const key in pathToValue) {
+    if (typeof pathToValue[key] === "function") {
+      _.updateWith(newObject, key, pathToValue[key]);
+    } else {
+      _.updateWith(newObject, key, (__) => pathToValue[key]);
+    }
+  }
+
+  return newObject;
 }
